Move ByteSlice to general utils
Also now can be logged as hex
