+++ internal/usbgadget/utils.go
@@ -2,7 +2,6 @@ package usbgadget
 
 import (
 	"bytes"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"os"
@@ -14,31 +13,6 @@ import (
 
 	"github.com/rs/zerolog"
 )
-
-type ByteSlice []byte
-
-func (s ByteSlice) MarshalJSON() ([]byte, error) {
-	vals := make([]int, len(s))
-	for i, v := range s {
-		vals[i] = int(v)
-	}
-	return json.Marshal(vals)
-}
-
-func (s *ByteSlice) UnmarshalJSON(data []byte) error {
-	var vals []int
-	if err := json.Unmarshal(data, &vals); err != nil {
-		return err
-	}
-	*s = make([]byte, len(vals))
-	for i, v := range vals {
-		if v < 0 || v > 255 {
-			return fmt.Errorf("value %d out of byte range", v)
-		}
-		(*s)[i] = byte(v)
-	}
-	return nil
-}
 
 func joinPath(basePath string, paths []string) string {
 	pathArr := append([]string{basePath}, paths...)
